requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
setuptools>=65.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
sentencepiece>=0.1.99